culling started, not fully working yet
Need to look at layers.  The collision sphere was colliding with the
tree's own leaves, so I had to add a new layer that had just the base
tree object and just ignore the leaves.  It's not working for some
reason, but might just need some other eyes on it.