revision7 - moved all prefabs to resources folder
scripts with a prefab reference will turn that reference into itself
so to fix that, I had the script record the prefab PATH rather than the
prefab itself. At runtime, it would find the prefab and store it as a
private variable. unforetunately, it means that all prefabs need to be
in the resources folder.