revision9 - almost done
system is in place! :D
i just need to add maxTrees and forest radius to biome script (right now
they are in TreeManager)

note that forests don't have special interactions with shrines yet.
trees should not findForest if they're inside of a shrine
and a forest should cancel creating a tree or seed if it realizes it's
trying to place one inside of a shrine (if the forest overlaps with a
shrine